Remove sudo module
OK, so I'd already done the work for sudo but forgot to add a NOPASSWD
thing. The work for librarian hasn't gone to waste for any future
changes but remove it for the time-being.

Also did the work to enable testing a simple Vagrant box with linking
the puppet repo in the Vagrantfile.